#27 [Fix] 보관함 - 내가 그린 코스 상세 페이지 이슈 해결
저번에 어떤 건 화면에 data가 잘 반영되고 어떤 건 안 된다고 했었는데 해결했음. 원인은 아래와 같음.

서버에서 distance 형식을 소수점으로 안 맞춰준 것 같음. 그래서 가령 1km 코스가 있다고 하면 이게 1.0으로 날아와야 하는데 1로 날아오니까 Dto를 만드는 과정에서 플러그인이 Float가 아닌 Int로 인식해버림.

그래서 2.4, 1.8과 같이 소수점이 있는 data가 서버에서 날아오면 수신은 잘 되는데 Dto가 잘못됐기 때문에 안드로이드 내에서 핸들링이 안 됐던 거였음. 

현재는 타입을 다 Float로 바꿔줌으로써 문제 해결했음.

나중에 기록 업로드 할 때 한 번 더 타입 맞춰줘야 함.

+++ app/src/main/java/com/runnect/runnect/data/model/DetailToRunData.kt
@@ -7,7 +7,7 @@ import kotlinx.android.parcel.Parcelize
 @Parcelize
 data class DetailToRunData(
     val departure: String,
-    val distance: Int,
+    val distance: Float,
     val path: List<List<Double>>,
     val image: String,
 ) : Parcelable
